move log_rec stuff & log_is_mvcc_op
