+++ src/transaction/log_manager.c
@@ -63,10 +63,11 @@
 #include "log_append.hpp"
 #include "log_archives.hpp"
 #include "log_compress.h"
-#include "partition_sr.h"
+#include "log_record.hpp"
 #include "log_system_tran.hpp"
 #include "log_volids.hpp"
 #include "log_writer.h"
+#include "partition_sr.h"
 #include "filter_pred_cache.h"
 #include "heap_file.h"
 #include "slotted_page.h"
