+++ src/transaction/log_recovery.c
@@ -37,6 +37,7 @@
 #include "log_impl.h"
 #include "log_lsa.hpp"
 #include "log_manager.h"
+#include "log_record.hpp"
 #include "log_system_tran.hpp"
 #include "log_volids.hpp"
 #include "recovery.h"
